+++ package.json
@@ -12,7 +12,8 @@
     "euler": "node index.js",
     "eulers/": "node index.js",
     "sync": "node scripts/sync_problems.js",
-    "usage": "node scripts/usage_leaderboard.js"
+    "usage": "node scripts/usage_leaderboard.js",
+    "new_lang": "node scripts/get_unused_langs.js"
   },
   "peerDependencies": {
     "typescript": "^5.0.0"
